Fixed ShapeAttributesTest.
The nested BalloonAttributeTests class was missing a call to
setMaximumSize. This resulted in a NullPointerException inside the
BasicBalloonAttributes.getRestorableState (on line 414). However, this
is probably not a very good fix. It allows the unit test to pass, but
some additional error handling needs to be added to the
BasicBalloonAttributes class to prevent this from happening.