Added pom.xml file.
- Added initial pom.xml file. Harvested some of the parts of this pom
  from this forum thread:
  http://forum.worldwindcentral.com/showthread.php?t=21381

Note: It will probably be best if this project is split into a couple of
maven sub-modules. The reason for this is that it will allow the
exclusion of the example classes from the final jar (almost like the
worldindx jar is currently built in a different manner when compared to
the standard worldwind jar).